Detect invalid rule name in command line option (updates #470)
This adds INFO level stats about total amount of rules and enabled
rules, makes `defaultRules` private, because they are not accessed
from anywhere. It also shows ERROR level message for invalid name
specified for --enable-rule or --disable-rule options.

The missing thing is to stop processing and with the error at this
point.