Call ApplyConfig before ValidateRules (#1043)
I'm working on a plugin which generates a list of rules dynamically,
based on its config:

https://github.com/richardTowers/tflint-ruleset-workspaces/

At the moment, it's not possible to refer to config in RuleNames because
that gets called before ApplyConfig. As a workaround, I've made my
plugin return the empty list for RuleNames:

https://github.com/richardTowers/tflint-ruleset-workspaces/blob/master/ruleset.go#L28

Naïvely, it looks like making the ApplyConfig a bit earlier would fix
this.

The tests pass, but I'm not sure if this will have unintended
consequences. Careful review would be much appreciated!