Introduce ephemeral input variables
Follow up of https://github.com/hashicorp/terraform/pull/35273
Follow up of https://github.com/hashicorp/terraform/pull/35985

Terraform throws an error if you use ephemeral values for the count meta-argument,
but TFLint does not. This is because the reason for throwing an error is that
plan cannot have ephemeral values, which is not an issue in the context of static analysis.
In the future, we can throw an error if we need to match this behavior.