Allow marked values in dynamic block `for_each`
Follow up of https://github.com/hashicorp/hcl/pull/679

Previously, for_each in dynamic blocks did not allow marked values
such as sensitive. However, https://github.com/hashicorp/hcl/pull/679
now supports this by propagating the marks to expanded children.

The reason behind this is to add a new mark called "ephemeral",
so we'll pull the changes to support Terraform 1.10.

Note that tfhcl's dynamic block support has incomplete mark propagation
since marked values resolve to unknown values. This is because in the past
the marked values could not be sent over the wire protocol,
and may be fixed in the near future.