Exclude expression context from marked `for_each` diagnostics (#2256)
Fixes https://github.com/terraform-linters/tflint/issues/2243
See also https://github.com/hashicorp/hcl/issues/737

The `hcl.DiagnosticTextWriter` has an issue where it will panic
when writing out a diagnostic that has an expression containing
a marked value.

TFLint can run into this issue when expanding the `for_each`
meta-argument with a sensitive value. If the `for_each` is not
iterable (it is not iterable if marked), a diagnostic containing
the expression context of the `for_each` is created and written
to stderr by the `DiagnosticTextWriter`.

Hopefully this will be fixed in upstream, but here's a workaround
to prevent the panic by excluding expression context from the
diagnostic when the `for_each` is marked.

By the way, other cases that include expression context in
diagnostics do not have the same issue, so no action is required.

- `for_each` is marked in dynamic blocks
  - The value is always unmarked, and non-iterable cases are rarer than those in meta-arguments
- dynamic block labels
  - Dynamic block labels are rarely used in Terraform
- `count` in meta-arguments
  - The value is always unmarked