Allow calling provider-defined functions with variadic arguments
Because TFLint cannot statically infer the behavior of provider-defined
functions, we took the approach of inferring the function interface
from function calls when parsing the HCL and generating mock functions.

As part of this, mocks were generated as fixed-length functions
based on the number of arguments, which caused errors when calling
provider-defined functions that accepted variadic argument,
such as `provider::deepmerge::merge`, with a different number of
arguments from the second time onwards.

This change fixes this issue by generating mock functions as always
accepting variadic arguments.